* Implement 16 bit PWM routines - Works on all BlinkM hardware * Make use of hardware timers and hardware PWM where appropriate * Use a lookup table to achieve linear brightness response during color mixing and fading. * Support BlinkM hardware using the attiny85 chips with 8k of flash
Although internally the PWM duty cycle has 16 bits of resolution, color values are still specified as 8 bits and then converted to an appropriate PWM value by the lookup table.

On the Attiny44/84 it is possible to use hardware PWM for all three colors; however there isn't any particular advantage in this situation as we don't really need the extra clock cycles for anything else. If there is a LOT of i2c traffic or something, this might become more important in the future.